+++ type/block.php
@@ -1,14 +1,92 @@
 <?php
+/**
+ * Contains the block class.
+ *
+ * @since 2.0.0
+ */
 
+defined( 'ABSPATH') || exit;
+
+/**
+ *
+ * The block super duper class.
+ *
+ *
+ * @since 2.0.0
+ * @version 2.0.0
+ */
 class WP_Super_Duper_Block {
 
-	public function __construct( $class ) {
-		$this->SD = $class;
-		$this->options = $class->options;
+	/**
+	 * @var WP_Super_Duper
+	 */
+	protected $sd;
 
+	/**
+	 * Class constructor.
+	 *
+	 * @param WP_Super_Duper $super_duper
+	 */
+	public function __construct( $super_duper ) {
+
+		$this->sd = $super_duper;
+
+		// Process the SD shortcode preview.
+		add_action( 'wp_ajax_super_duper_output_shortcode', array( __CLASS__, 'render_shortcode' ) );
 		add_action( 'admin_enqueue_scripts', array( $this, 'register_block' ) );
 		add_action( 'admin_footer-post.php', array( $this,'enqueue_editor_assets' ) );
 		add_action( 'admin_footer-post-new.php', array( $this,'enqueue_editor_assets' ) );
+	}
+
+	/**
+	 * Render the shortcode via ajax so we can return it to Gutenberg.
+	 *
+	 * @since 1.0.0
+	 */
+	public static function render_shortcode() {
+
+		// Verify nonce.
+		check_ajax_referer( 'super_duper_output_shortcode', '_ajax_nonce', true );
+
+		// And permissions.
+		if ( ! current_user_can( 'manage_options' ) ) {
+			wp_die();
+		}
+
+		// Ensure a shortcode is provided.
+		if ( empty( $_POST['shortcode'] ) ) {
+			wp_die();
+		}
+
+		// We might need the $post value here so lets set it.
+		if ( isset( $_POST['post_id'] ) && $_POST['post_id'] ) {
+			$post_obj = get_post( absint( $_POST['post_id'] ) );
+			if ( ! empty( $post_obj ) && empty( $post ) ) {
+				global $post;
+				$post = $post_obj;
+			}
+		}
+
+		// Prepare args.
+		$shortcode_name   = sanitize_title_with_dashes( $_POST['shortcode'] );
+		$attributes_array = isset( $_POST['attributes'] ) && $_POST['attributes'] ? $_POST['attributes'] : array();
+		$attributes       = '';
+
+		// Convert attributes array into a string.
+		if ( ! empty( $attributes_array ) ) {
+			foreach ( $attributes_array as $key => $value ) {
+				if ( is_array( $value ) ) {
+					$value = implode( ',', $value );
+				}
+				$key         = sanitize_title_with_dashes( $key );
+				$value       = wp_slash( $value );
+				$attributes .= " $key='$value' ";
+			}
+		}
+
+		echo do_shortcode( "[$shortcode_name $attributes]" );
+		exit;
+
 	}
 
 	/**
@@ -17,73 +95,8 @@ class WP_Super_Duper_Block {
 	public function register_block() {
 		wp_add_inline_script( 'wp-blocks', $this->block() );
 		if ( class_exists( 'SiteOrigin_Panels' ) ) {
-			wp_add_inline_script( 'wp-blocks', $this->siteorigin_js() );
+			wp_add_inline_script( 'wp-blocks', $this->sd->siteorigin_js() );
 		}
-	}
-
-	/**
-	 * Enqueue CSS assets in the editor if defined
-	 */
-	public function enqueue_editor_assets() {
-		if ( isset( $this->options['block-editor-style'] ) ) {
-			echo "<link rel='stylesheet' id='glossary-shortcode' href='" . $this->options['block-editor-style'] . "'/>";
-		}
-	}
-
-	/**
-	 * Check if we need to show advanced options.
-	 *
-	 * @return bool
-	 */
-	public function block_show_advanced() {
-		$show      = false;
-		$this->arguments = $this->SD->get_arguments();
-		$arguments = $this->arguments;
-
-		if ( ! empty( $arguments ) ) {
-			foreach ( $arguments as $argument ) {
-				if ( isset( $argument['advanced'] ) && $argument['advanced'] ) {
-					$show = true;
-					break; // no need to continue if we know we have it
-				}
-			}
-		}
-
-		return $show;
-	}
-
-	/**
-	 * Generate the block icon.
-	 *
-	 * Enables the use of Font Awesome icons.
-	 *
-	 * @note xlink:href is actually deprecated but href is not supported by all so we use both.
-	 *
-	 * @param $icon
-	 *
-	 * @since 1.1.0
-	 * @return string
-	 */
-	public function get_block_icon( $icon ) {
-		// check if we have a Font Awesome icon
-		$fa_type = '';
-		if ( substr( $icon, 0, 7 ) === "fas fa-" ) {
-			$fa_type = 'solid';
-		} elseif ( substr( $icon, 0, 7 ) === "far fa-" ) {
-			$fa_type = 'regular';
-		} elseif ( substr( $icon, 0, 7 ) === "fab fa-" ) {
-			$fa_type = 'brands';
-		} else {
-			$icon = "'" . $icon . "'";
-		}
-
-		// set the icon if we found one
-		if ( $fa_type ) {
-			$fa_icon = str_replace( array( "fas fa-", "far fa-", "fab fa-" ), "", $icon );
-			$icon    = "el('svg',{width: 20, height: 20, viewBox: '0 0 20 20'},el('use', {'xlink:href': '" . $this->SD->get_url() . "icons/" . $fa_type . ".svg#" . $fa_icon . "','href': '" . $this->SD->get_url() . "icons/" . $fa_type . ".svg#" . $fa_icon . "'}))";
-		}
-
-		return $icon;
 	}
 
 	/**
@@ -95,146 +108,147 @@ class WP_Super_Duper_Block {
 	 * @return mixed
 	 */
 	public function block() {
+
 		ob_start();
+		$show_advanced        = $this->sd->block_show_advanced();
+		$js_id                = str_replace( '-', '_', $this->sd->base_id );
+		$post_type_rest_slugs = array();
 
-		$show_advanced = $this->block_show_advanced();
+		if ( ! empty( $this->sd->arguments ) && isset( $this->sd->arguments['post_type']['onchange_rest']['values'] ) ) {
+			$post_type_rest_slugs = array( $this->sd->arguments['post_type']['onchange_rest']['values'] );
+		}
+
 		?>
-		<script>
-			/**
-			 * BLOCK: Basic
-			 *
-			 * Registering a basic block with Gutenberg.
-			 * Simple block, renders and saves the same
-			 /without any interactivity.
-			 *
-			 * Styles:
-			 *        editor.css — Editor styles for the block.
-			 *        style.css  — Editor & Front end styles for the block.
-			 */
-			(function () {
-				var __ = wp.i18n.__; // The __() for internationalization.
-				var el = wp.element.createElement; // The wp.element.createElement() function to create elements.
-				var editable = wp.blocks.Editable;
-				var blocks = wp.blocks;
-				var registerBlockType = wp.blocks.registerBlockType; // The registerBlockType() to register blocks.
-				var is_fetching = false;
-				var prev_attributes = [];
-
-				var term_query_type = '';
-				var post_type_rest_slugs = <?php if(! empty( $this->arguments ) && isset($this->arguments['post_type']['onchange_rest']['values'])){echo "[".json_encode($this->arguments['post_type']['onchange_rest']['values'])."]";}else{echo "[]";} ?>;
-				const taxonomies_<?php echo str_replace("-","_", $this->SD->base_id);?> = [{label: "Please wait", value: 0}];
-				const sort_by_<?php echo str_replace("-","_", $this->SD->base_id);?> = [{label: "Please wait", value: 0}];
-
+			<script>
 				/**
-				 * Register Basic Block.
+				 * BLOCK: Basic
 				 *
-				 * Registers a new block provided a unique name and an object defining its
-				 * behavior. Once registered, the block is made available as an option to any
-				 * editor interface where blocks are implemented.
+				 * Registering a basic block with Gutenberg.
+				 * Simple block, renders and saves the same content without any interactivity.
 				 *
-				 * @param  {string}   name     Block name.
-				 * @param  {Object}   settings Block settings.
-				 * @return {?WPBlock}          The block, if it has been successfully
-				 *                             registered; otherwise `undefined`.
+				 * Styles:
+				 *        editor.css — Editor styles for the block.
+				 *        style.css  — Editor & Front end styles for the block.
 				 */
-				registerBlockType('<?php echo str_replace( "_", "-", sanitize_title_with_dashes( $this->options['textdomain'] ) . '/' . sanitize_title_with_dashes( $this->options['class_name'] ) );  ?>', { // Block name. Block names must be string that contains a namespace prefix. Example: my-plugin/my-custom-block.
-					title: '<?php echo addslashes( $this->options['name'] ); ?>', // Block title.
-					description: '<?php echo addslashes( $this->options['widget_ops']['description'] )?>', // Block title.
-					icon: <?php echo $this->get_block_icon( $this->options['block-icon'] );?>,//'<?php echo isset( $this->options['block-icon'] ) ? esc_attr( $this->options['block-icon'] ) : 'shield-alt';?>', // Block icon from Dashicons → https://developer.wordpress.org/resource/dashicons/.
-					supports: {
+				(function () {
+					var __ = wp.i18n.__; // The __() for internationalization.
+					var el = wp.element.createElement; // The wp.element.createElement() function to create elements.
+					var editable = wp.blocks.Editable;
+					var blocks = wp.blocks;
+					var registerBlockType = wp.blocks.registerBlockType; // The registerBlockType() to register blocks.
+					var is_fetching = false;
+					var prev_attributes = [];
+
+					var term_query_type = '';
+					var post_type_rest_slugs = <?php json_encode( $post_type_rest_slugs ); ?>;
+					const taxonomies_<?php echo esc_js( $js_id ); ?> = [{label: __( "Please wait", '<?php echo esc_js( $this->sd->options['textdomain'] ); ?>' ), value: 0}];
+					const sort_by_<?php echo esc_js( $js_id ); ?> = [{label: __( "Please wait", '<?php echo esc_js( $this->sd->options['textdomain'] ); ?>' ), value: 0}];
+
+					/**
+					 * Register Basic Block.
+					 *
+					 * Registers a new block provided a unique name and an object defining its
+					 * behavior. Once registered, the block is made available as an option to any
+					 * editor interface where blocks are implemented.
+					 *
+					 * @param  {string}   name     Block name.
+					 * @param  {Object}   settings Block settings.
+					 * @return {?WPBlock}          The block, if it has been successfully
+					 *                             registered; otherwise `undefined`.
+					 */
+					registerBlockType('<?php echo esc_js( str_replace( '_', '-', sanitize_title_with_dashes( $this->sd->options['textdomain'] ) ) . '/' . esc_js( sanitize_title_with_dashes( $this->sd->options['class_name'] ) ) );  ?>', { // Block name. Block names must be string that contains a namespace prefix. Example: my-plugin/my-custom-block.
+						title: '<?php echo addslashes( $this->sd->options['name'] ); ?>', // Block title.
+						description: '<?php echo addslashes( $this->sd->options['widget_ops']['description'] )?>', // Block description.
+						icon: <?php echo $this->get_block_icon( $this->sd->options['block-icon'] ); ?>,//'<?php echo isset( $this->sd->options['block-icon'] ) ? esc_attr( $this->sd->options['block-icon'] ) : 'shield-alt';?>', // Block icon from Dashicons → https://developer.wordpress.org/resource/dashicons/.
+						supports: { <?php echo isset( $this->sd->options['block-supports'] ) ? $this->sd->array_to_attributes( $this->sd->options['block-supports'] ) : '' ?> },
+						category: '<?php echo isset( $this->sd->options['block-category'] ) ? esc_js( $this->sd->options['block-category'] ) : 'common';?>', // Block category — Group blocks together based on common traits E.g. common, formatting, layout widgets, embed.
+						<?php if ( isset( $this->sd->options['block-keywords'] ) ) :
+							printf( 'keywords : %s,', $this->sd->options['block-keywords'] );
+						endif; ?>
+
 						<?php
-						if ( isset( $this->options['block-supports'] ) ) {
-							echo $this->array_to_attributes( $this->options['block-supports'] );
-						}
-						?>
-					},
-					category: '<?php echo isset( $this->options['block-category'] ) ? esc_attr( $this->options['block-category'] ) : 'common';?>', // Block category — Group blocks together based on common traits E.g. common, formatting, layout widgets, embed.
-					<?php if ( isset( $this->options['block-keywords'] ) ) {
-						echo "keywords: " . $this->options['block-keywords'] . ",\n";
-					}
-					// maybe set no_wrap
-					$no_wrap = isset( $this->options['no_wrap'] ) && $this->options['no_wrap'] ? true : false;
-					if ( isset( $this->arguments['no_wrap'] ) && $this->arguments['no_wrap'] ) {
-						$no_wrap = true;
-					}
-					if ( $no_wrap ) {
-						$this->options['block-wrap'] = '';
-					}
 
-					$show_alignment = false;
-					// align feature
-					/*echo "supports: {";
-					echo "	align: true,";
-					echo "  html: false";
-					echo "},";*/
-
-					if ( ! empty( $this->arguments ) ) {
-						echo "attributes: {\n";
-						if ( $show_advanced ) {
-							echo "show_advanced: {";
-							echo "type: 'boolean',";
-							echo " default: false";
-							echo "},\n";
-						}
-
-						// block wrap element
-						if ( ! empty( $this->options['block-wrap'] ) ) { //@todo we should validate this?
-							echo "block_wrap: {";
-							echo "type: 'string',";
-							echo " default: '" . esc_attr( $this->options['block-wrap'] ) . "'";
-							echo "},\n";
-						}
-
-						foreach ( $this->arguments as $key => $args ) {
-							// set if we should show alignment
-							if ( $key == 'alignment' ) {
-								$show_alignment = true;
+							// maybe set no_wrap
+							if ( ! empty( $this->sd->options['no_wrap'] ) || ! empty( $this->sd->arguments['no_wrap'] ) ) {
+								$this->sd->options['block-wrap'] = '';
 							}
 
-							if ( $args['type'] == 'checkbox' ) {
-								$type    = 'boolean';
-								$default = isset( $args['default'] ) && $args['default'] ? 'true' : 'false';
-							} elseif ( $args['type'] == 'number' ) {
-								$type    = 'number';
-								$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
-							} elseif ( $args['type'] == 'select' && ! empty( $args['multiple'] ) ) {
-								$type = 'array';
-								if ( is_array( $args['default'] ) ) {
-									$default = isset( $args['default'] ) ? "['" . implode( "','", $args['default'] ) . "']" : "[]";
-								} else {
-									$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
+							$show_alignment = false;
+
+							if ( ! empty( $this->sd->arguments ) ) {
+								echo "attributes : {";
+
+								if ( $show_advanced ) {
+									echo "show_advanced: {";
+									echo "	type: 'boolean',";
+									echo "  default: false,";
+									echo "},";
 								}
-							} elseif ( $args['type'] == 'multiselect' ) {
-								$type    = 'array';
-								$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
-							} else {
-								$type    = 'string';
-								$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
+
+								// block wrap element
+								if ( ! empty( $this->sd->options['block-wrap'] ) ) { //@todo we should validate this?
+									echo "block_wrap: {";
+									echo "	type: 'string',";
+									echo "  default: '" . esc_js( $this->sd->options['block-wrap'] ) . "',";
+									echo "},";
+								}
+
+								foreach ( $this->sd->arguments as $key => $args ) {
+
+									// set if we should show alignment
+									if ( $key == 'alignment' ) {
+										$show_alignment = true;
+									}
+
+									$extra = '';
+
+									if ( $args['type'] == 'checkbox' ) {
+										$type    = 'boolean';
+										$default = isset( $args['default'] ) && $args['default'] ? 'true' : 'false';
+									} elseif ( $args['type'] == 'number' ) {
+										$type    = 'number';
+										$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
+									} elseif ( $args['type'] == 'select' && ! empty( $args['multiple'] ) ) {
+										$type = 'array';
+										if ( isset( $args['default'] ) && is_array( $args['default'] ) ) {
+											$default = ! empty( $args['default'] ) ? "['" . implode( "','", $args['default'] ) . "']" : "[]";
+										} else {
+											$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
+										}
+									} elseif ( $args['type'] == 'multiselect' ) {
+										$type    = 'array';
+										$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
+									} else {
+										$type    = 'string';
+										$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
+									}
+									echo $key . " : {";
+									echo "type : '$type',";
+									echo "default : $default,";
+									echo "},";
+								}
+
+								echo "content : {type : 'string',default: 'Please select the attributes in the block settings'},";
+								echo "className: { type: 'string', default: '' },";
+
+								echo "},";
+
 							}
-							echo str_replace( '-','__', $key ) . ": {";
-							echo "type: '$type',";
-							echo "default: $default";
-							echo "},\n";
-						}
 
-						echo "content : {type : 'string',default: 'Please select the attributes in the block settings'},\n";
-						echo "className: { type: 'string', default: '' },\n";
-						echo "},";
-
-					}
-				?>
-
-				// The "edit" property must be a valid function.
-					edit: function (props) {
-						var $value = '';
-					<?php
-						// if we have a post_type and a category then link them
-						if( isset($this->arguments['post_type']) && isset($this->arguments['category']) && !empty($this->arguments['category']['post_type_linked']) ){
 						?>
-						if(typeof(prev_attributes[props.id]) != 'undefined' ){
-							$pt = props.attributes.post_type;
-							if(post_type_rest_slugs.length){
-								$value = post_type_rest_slugs[0][$pt];
+
+						// The "edit" property must be a valid function.
+						edit: function (props) {
+
+							var $value = '';
+							<?php
+								// if we have a post_type and a category then link them
+								if( isset( $this->sd->arguments['post_type'] ) && isset( $this->sd->arguments['category'] ) && !empty($this->sd->arguments['category']['post_type_linked']) ){
+							?>
+							if(typeof(prev_attributes[props.id]) != 'undefined' ){
+								$pt = props.attributes.post_type;
+								if(post_type_rest_slugs.length){
+									$value = post_type_rest_slugs[0][$pt];
 								}
 								var run = false;
 
@@ -245,7 +259,7 @@ class WP_Super_Duper_Block {
 
 								// taxonomies
 								if( $value && 'post_type' in prev_attributes[props.id] && 'category' in prev_attributes[props.id] && run ){
-									wp.apiFetch({path: "<?php if(isset($this->arguments['post_type']['onchange_rest']['path'])){echo $this->arguments['post_type']['onchange_rest']['path'];}else{'/wp/v2/"+$value+"/categories';} ?>"}).then(terms => {
+									wp.apiFetch({path: "<?php if(isset($this->sd->arguments['post_type']['onchange_rest']['path'])){echo $this->sd->arguments['post_type']['onchange_rest']['path'];}else{'/wp/v2/"+$value+"/categories/?per_page=100';} ?>"}).then(terms => {
 										while (taxonomies_<?php echo str_replace("-","_", $this->id);?>.length) {
 										taxonomies_<?php echo str_replace("-","_", $this->id);?>.pop();
 									}
@@ -289,104 +303,116 @@ class WP_Super_Duper_Block {
 
 								}
 							}
-						<?php }?>
+							<?php }?>
 
-						var content = props.attributes.content;
 
-						function onChangeContent() {
-							$refresh = false;
+							var content = props.attributes.content;
 
-							// Set the old content the same as the new one so we only compare all other attributes
-							if(typeof(prev_attributes[props.id]) != 'undefined'){
-								prev_attributes[props.id].content = props.attributes.content;
-							}else if(props.attributes.content === ""){
-								// if first load and content empty then refresh
-								$refresh = true;
+							function onChangeContent() {
+
+								$refresh = false;
+
+								// Set the old content the same as the new one so we only compare all other attributes
+								if(typeof(prev_attributes[props.id]) != 'undefined'){
+									prev_attributes[props.id].content = props.attributes.content;
+								}else if(props.attributes.content === ""){
+									// if first load and content empty then refresh
+									$refresh = true;
+								}
+
+								if ( ( !is_fetching &&  JSON.stringify(prev_attributes[props.id]) != JSON.stringify(props.attributes) ) || $refresh  ) {
+
+									is_fetching = true;
+									var data = {
+										'action': 'super_duper_output_shortcode',
+										'shortcode': '<?php echo $this->sd->options['base_id'];?>',
+										'attributes': props.attributes,
+										'post_id': <?php global $post; if ( isset( $post->ID ) ) {
+										echo $post->ID;
+									}else{echo '0';}?>,
+										'_ajax_nonce': '<?php echo wp_create_nonce( 'super_duper_output_shortcode' );?>'
+									};
+
+									jQuery.post(ajaxurl, data, function (response) {
+										return response;
+									}).then(function (env) {
+
+										// if the content is empty then we place some placeholder text
+										if (env == '') {
+											env = "<div style='background:#0185ba33;padding: 10px;border: 4px #ccc dashed;'>" + "<?php _e( 'Placeholder for: ' );?>" + props.name + "</div>";
+										}
+
+										props.setAttributes({content: env});
+										is_fetching = false;
+										prev_attributes[props.id] = props.attributes;
+
+										// if AUI is active call the js init function
+										if (typeof aui_init === "function") {
+											aui_init();
+										}
+									});
+
+
+								}
+
+								return props.attributes.content;
+
 							}
 
-							if ( ( !is_fetching &&  JSON.stringify(prev_attributes[props.id]) != JSON.stringify(props.attributes) ) || $refresh  ) {
-								is_fetching = true;
-								var data = {
-									'action': 'super_duper_output_shortcode',
-									'shortcode': '<?php echo $this->options['base_id'];?>',
-									'attributes': props.attributes,
-									'post_id': <?php global $post; if ( isset( $post->ID ) ) {
-									echo $post->ID;
-								}else{echo '0';}?>,
-									'_ajax_nonce': '<?php echo wp_create_nonce( 'super_duper_output_shortcode' );?>'
-								};
+							return [
 
-								jQuery.post(ajaxurl, data, function (response) {
-									return response;
-								}).then(function (env) {
-									// if the content is empty then we place some placeholder text
-									if (env == '') {
-										env = "<div style='background:#0185ba33;padding: 10px;border: 4px #ccc dashed;'>" + "<?php _e( 'Placeholder for: ' );?>" + props.name + "</div>";
-									}
+								el(wp.blockEditor.BlockControls, {key: 'controls'},
 
-									props.setAttributes({content: env});
-									is_fetching = false;
-									prev_attributes[props.id] = props.attributes;
-
-									// if AUI is active call the js init function
-									if (typeof aui_init === "function") {
-										aui_init();
-									}
-								});
-
-							}
-
-							return props.attributes.content;
-
-						}
-
-						return [
-							el(wp.blockEditor.BlockControls, {key: 'controls'},
-
-								<?php if($show_alignment){?>
-								el(
-									wp.blockEditor.AlignmentToolbar,
-									{
-										value: props.attributes.alignment,
-										onChange: function (alignment) {
-											props.setAttributes({alignment: alignment})
-										},
-										key: props.attributes.alignment
-									}
-								)
-								<?php }?>
-							),
-							el(wp.blockEditor.InspectorControls, {key: 'inspector'},
-								<?php
-									if(! empty( $this->arguments )){
-
-									if ( $show_advanced ) {
-								?>
-								el('div', {
-										style: {'paddingLeft': '16px','paddingRight': '16px'}
-									},
+									<?php if($show_alignment){?>
 									el(
-										wp.components.ToggleControl,
+										wp.blockEditor.AlignmentToolbar,
 										{
-											label: 'Show Advanced Settings?',
-											checked: props.attributes.show_advanced,
-											onChange: function (show_advanced) {
-												props.setAttributes({show_advanced: !props.attributes.show_advanced})
-											},
-											key: props.attributes.show_advanced
+											value: props.attributes.alignment,
+											onChange: function (alignment) {
+												props.setAttributes({alignment: alignment})
+											}
 										}
 									)
+									<?php }?>
+
 								),
-								<?php
+
+								el(wp.blockEditor.InspectorControls, {key: 'inspector'},
+
+									<?php
+
+									if(! empty( $this->sd->arguments )){
+
+									if ( $show_advanced ) {
+									?>
+									el('div', {
+											style: {'padding-left': '16px','padding-right': '16px'}
+										},
+										el(
+											wp.components.ToggleControl,
+											{
+												label: 'Show Advanced Settings?',
+												checked: props.attributes.show_advanced,
+												onChange: function (show_advanced) {
+													props.setAttributes({show_advanced: !props.attributes.show_advanced})
+												}
+											}
+										)
+									)
+									,
+									<?php
+
 									}
 
-									$arguments = $this->SD->group_arguments( $this->arguments );
+									$arguments = $this->sd->group_arguments( $this->sd->arguments );
+
 									// Do we have sections?
-									$has_sections = $arguments == $this->arguments ? false : true;
+									$has_sections = $arguments == $this->sd->arguments ? false : true;
+
 
 									if($has_sections){
-										$panel_count = 0;
-										foreach($arguments as $key => $args){
+									$panel_count = 0;
+									foreach($arguments as $key => $args){
 									?>
 									el(wp.components.PanelBody, {
 											title: '<?php esc_attr_e( $key ); ?>',
@@ -397,8 +423,11 @@ class WP_Super_Duper_Block {
 										}?>
 										},
 										<?php
+
+
+
 										foreach ( $args as $k => $a ) {
-                                            $k = str_replace('-','__', $k);
+
 											$this->block_row_start( $k, $a );
 											$this->build_block_arguments( $k, $a );
 											$this->block_row_end( $k, $a );
@@ -407,41 +436,41 @@ class WP_Super_Duper_Block {
 									),
 									<?php
 									$panel_count ++;
-										}
-									} else {
+
+									}
+									}else {
 									?>
 									el(wp.components.PanelBody, {
-										title: '<?php esc_attr_e( "Settings" ); ?>',
-										initialOpen: true
-									},
-									<?php
-										foreach ( $this->arguments as $key => $args ) {
-                                            $key = str_replace('-','__', $key);
+											title: '<?php esc_attr_e( "Settings" ); ?>',
+											initialOpen: true
+										},
+										<?php
+										foreach ( $this->sd->arguments as $key => $args ) {
 											$this->block_row_start( $key, $args );
 											$this->build_block_arguments( $key, $args );
 											$this->block_row_end( $key, $args );
 										}
-									?>
+										?>
 									),
 									<?php
 									}
 
-								}
-							?>
-							),
+									}
+									?>
 
-							<?php
-							// If the user sets block-output array then build it
-							if ( ! empty( $this->options['block-output'] ) ) {
-								$this->block_element( $this->options['block-output'] );
-							} else {
+								),
+
+								<?php
+								// If the user sets block-output array then build it
+								if ( ! empty( $this->sd->options['block-output'] ) ) {
+								$this->block_element( $this->sd->options['block-output'] );
+							}else{
 								// if no block-output is set then we try and get the shortcode html output via ajax.
 								?>
 								el('div', {
 									dangerouslySetInnerHTML: {__html: onChangeContent()},
 									className: props.className,
-									style: {'minHeight': '30px'},
-									key: props.className + '-output'
+									style: {'minHeight': '30px'}
 								})
 								<?php
 								}
@@ -451,27 +480,30 @@ class WP_Super_Duper_Block {
 
 						// The "save" property must be specified and must be a valid function.
 						save: function (props) {
+
+							//console.log(props);
+
+
 							var attr = props.attributes;
 							var align = '';
 
 							// build the shortcode.
-							var content = "[<?php echo $this->options['base_id'];?>";
+							var content = "[<?php echo $this->sd->options['base_id'];?>";
 							$html = '';
 							<?php
 
-							if(! empty( $this->arguments )){
+							if(! empty( $this->sd->arguments )){
 
-							foreach($this->arguments as $key => $args){
-								$key = str_replace('-','__', $key);
-								?>
-								if (attr.hasOwnProperty("<?php echo esc_attr( $key );?>")) {
-									if ('<?php echo esc_attr( $key );?>' == 'html') {
-										$html = attr.<?php echo esc_attr( $key );?>;
-									} else {
-										content += " <?php echo str_replace( '__','-', esc_attr( $key ) );?>='" + attr.<?php echo esc_attr( $key );?>+ "' ";
-									}
+							foreach($this->sd->arguments as $key => $args){
+							?>
+							if (attr.hasOwnProperty("<?php echo esc_attr( $key );?>")) {
+								if ('<?php echo esc_attr( $key );?>' == 'html') {
+									$html = attr.<?php echo esc_attr( $key );?>;
+								} else {
+									content += " <?php echo esc_attr( $key );?>='" + attr.<?php echo esc_attr( $key );?>+ "' ";
 								}
-								<?php
+							}
+							<?php
 							}
 							}
 
@@ -480,8 +512,9 @@ class WP_Super_Duper_Block {
 
 							// if has html element
 							if ($html) {
-								content += $html + "[/<?php echo $this->options['base_id'];?>]";
+								content += $html + "[/<?php echo $this->sd->options['base_id'];?>]";
 							}
+
 
 							// @todo should we add inline style here or just css classes?
 							if (attr.alignment) {
@@ -497,36 +530,86 @@ class WP_Super_Duper_Block {
 							}
 
 							<?php
-								if(isset( $this->options['block-wrap'] ) && $this->options['block-wrap'] == ''){
+							if(isset( $this->sd->options['block-wrap'] ) && $this->sd->options['block-wrap'] == ''){
 							?>
 							return content;
 							<?php
-							} else {
+							}else{
 							?>
-								var block_wrap = 'div';
-								if (attr.hasOwnProperty("block_wrap")) {
-									block_wrap = attr.block_wrap;
-								}
-								return el(block_wrap, {dangerouslySetInnerHTML: {__html: content}, className: align});
-								<?php
+							var block_wrap = 'div';
+							if (attr.hasOwnProperty("block_wrap")) {
+								block_wrap = attr.block_wrap;
+							}
+							return el(block_wrap, {dangerouslySetInnerHTML: {__html: content}, className: align});
+							<?php
 							}
 							?>
+
 
 						}
 					});
 				})();
 			</script>
-		<?php
-		$output = ob_get_clean();
+			<?php
+			$output = ob_get_clean();
 
-		/*
-		 * We only add the <script> tags for code highlighting, so we strip them from the output.
-		 */
+			/*
+			 * We only add the <script> tags for code highlighting, so we strip them from the output.
+			 */
 
-		return str_replace( array(
-			'<script>',
-			'</script>'
-		), '', $output );
+			return str_replace(
+				array(
+					'<script>',
+					'</script>'
+				),
+				'',
+				$output
+			);
+
+		}
+
+	/**
+	 * Generate the block icon.
+	 *
+	 * Enables the use of Font Awesome icons.
+	 *
+	 * @note xlink:href is actually deprecated but href is not supported by all so we use both.
+	 *
+	 * @param $icon
+	 *
+	 * @since 1.1.0
+	 * @return string
+	 */
+	public function get_block_icon( $icon ) {
+
+		// check if we have a Font Awesome icon
+		$fa_type = '';
+		if ( substr( $icon, 0, 7 ) === 'fas fa-' ) {
+			$fa_type = 'solid';
+		} elseif ( substr( $icon, 0, 7 ) === 'far fa-' ) {
+			$fa_type = 'regular';
+		} elseif ( substr( $icon, 0, 7 ) === 'fab fa-' ) {
+			$fa_type = 'brands';
+		} else {
+			$icon = sprintf( "'%s'", esc_js( $icon ) );
+		}
+
+		// set the icon if we found one
+		if ( $fa_type ) {
+			$fa_icon =  substr( $icon, 7 );
+			$icon    = "el('svg',{width: 20, height: 20, viewBox: '0 0 20 20'},el('use', {'xlink:href': '" . $this->sd->get_url() . "icons/" . $fa_type . ".svg#" . $fa_icon . "','href': '" . $this->sd->get_url() . "icons/" . $fa_type . ".svg#" . $fa_icon . "'}))";
+		}
+
+		return $icon;
+	}
+
+	/**
+	 * Enqueue CSS assets in the editor if defined
+	 */
+	public function enqueue_editor_assets() {
+		if ( isset( $this->sd->options['block-editor-style'] ) ) {
+			echo "<link rel='stylesheet' id='glossary-shortcode' href='" . $this->sd->options['block-editor-style'] . "'/>";
+		}
 	}
 
 	public function block_row_start($key, $args){
@@ -601,7 +684,7 @@ class WP_Super_Duper_Block {
 	}
 
 	public function build_block_arguments( $key, $args ) {
-		$custom_attributes = ! empty( $args['custom_attributes'] ) ? $this->SD->array_to_attributes( $args['custom_attributes'] ) : '';
+		$custom_attributes = ! empty( $args['custom_attributes'] ) ? $this->sd->array_to_attributes( $args['custom_attributes'] ) : '';
 		$options           = '';
 		$extra             = 'key: \'' . $key . '\',' . "\n";
 
@@ -614,7 +697,7 @@ class WP_Super_Duper_Block {
 		$icon = '';
 		if( !empty( $args['icon'] ) ){
 			$icon .= "el('div', {";
-								$icon .= "dangerouslySetInnerHTML: {__html: '".$this->SD->get_widget_icon( esc_attr($args['icon']))."'},";
+								$icon .= "dangerouslySetInnerHTML: {__html: '".$this->sd->get_widget_icon( esc_attr($args['icon']))."'},";
 								$icon .= "className: 'text-center',";
 								$icon .= "title: '".addslashes( $args['title'] )."',";
 							$icon .= "}),";
@@ -720,7 +803,6 @@ class WP_Super_Duper_Block {
 		<?php
 	}
 
-
 	/**
 	 * A self looping function to create the output for JS block elements.
 	 *
@@ -729,9 +811,15 @@ class WP_Super_Duper_Block {
 	 * @param $args
 	 */
 	public function block_element( $args ) {
+
+		// Make sure we have args.
 		if ( ! empty( $args ) ) {
+
+			// For each argument...
 			foreach ( $args as $element => $new_args ) {
-				if ( is_array( $new_args ) ) { // its an element
+
+				// Check if its an element.
+				if ( is_array( $new_args ) ) {
 					if ( isset( $new_args['element'] ) ) {
 						if ( isset( $new_args['element_require'] ) ) {
 							echo str_replace( array(
@@ -786,9 +874,11 @@ class WP_Super_Duper_Block {
 						echo ")";// end content
 						echo ", \n";
 					}
+
+				// Or an attribute.
 				} else {
-					if ( substr( $element, 0, 3 ) === "if_" ) {
-						echo str_replace( "if_", "", $element ) . ": " . $this->block_props_replace( $new_args, true ) . ",";
+					if ( substr( $element, 0, 3 ) === 'if_' ) {
+						echo str_replace( 'if_', '', $element ) . ": " . $this->block_props_replace( $new_args, true ) . ",";
 					} elseif ( $element == 'style' ) {
 						echo $element . ": " . $this->block_props_replace( $new_args ) . ",";
 					} else {
@@ -796,6 +886,7 @@ class WP_Super_Duper_Block {
 					}
 				}
 			}
+
 		}
 	}
 
@@ -807,6 +898,7 @@ class WP_Super_Duper_Block {
 	 * @return mixed
 	 */
 	public function block_props_replace( $string, $no_wrap = false ) {
+
 		if ( $no_wrap ) {
 			$string = str_replace( array( "[%", "%]" ), array( "props.attributes.", "" ), $string );
 		} else {
